1. pending error donate api and  join can join donate or not trigger

+++ app/src/main/java/mithril/hackathon/chasingcoin/data/network/server/response/StatsResp.kt
@@ -28,11 +28,11 @@ data class StatsResp(
 
     data class My(
         @SerializedName("amount")
-        val amount: Int,
+        val amount: Long,
         @SerializedName("rank")
         val rank: String,
         @SerializedName("totalDistance")
-        val totalDistance: Int
+        val totalDistance: Double
     )
